last minute fixes and refactoring
