+++ BlazorClient/Pages/EditProduct.razor
@@ -50,7 +50,7 @@ else
     {
         try
         {
-            product = await Http.GetFromJsonAsync<Product>($"https://localhost:7005/products/{ProductId}");
+            product = await Http.GetFromJsonAsync<Product>($"/products/{ProductId}");
         }
         catch(Exception ex)
         {
@@ -60,7 +60,7 @@ else
 
     private async Task HandleValidSubmit()
     {
-        var response = await Http.PutAsJsonAsync($"https://localhost:7005/products/{ProductId}", product);
+        var response = await Http.PutAsJsonAsync($"/products/{ProductId}", product);
         if (response.IsSuccessStatusCode)
         {
             Navigation.NavigateTo("/fetchproducts");
